Fix error when raw output from jq produces empty string
Empty string causes json.Unmarshal to fail within PrettyPrint. It would
fail on encoder.UnmarshalJSONBytes but the JSON encoder no-ops because
it assumes the jq output isn't empty.

Since the jq output is empty, I opted to assuming all encoders should
output nothing in this case and checked if the program output is empty
in printValue, rather than in jsonEncoding.PrettyPrint.